requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
strands-agents
strands-agents-tools